Various updates to bulktester and cachecleaner. Now recommending only using file store for caching.

+++ classes/task/cache_cleaner.php
@@ -57,45 +57,51 @@ class cache_cleaner extends \core\task\scheduled_task {
     }
 
     /**
-     * Execute the task.
+     * This task will purge old cache entries in file caches used
+     * by the Coderunner grading cache definition.
+     * Task does nothing if the grading cache isn't enabled
+     * in the Coderunner admin settings. You might want
+     * to purge the Coderunner grading cache manually
+     * if you aren't going to turn it on again.
      */
     public function execute() {
-        // Use system context as purging everything
-        // ... and set use TTL to true so that only old
-        // ... entries are purged.
-        $purger = new cache_purger(1, true);
-        $definition = $purger->get_coderunner_cache_definition();
-        $store = $purger->get_first_file_store($definition);
-        $ttl = $purger->ttl;
-        if ($ttl) {
-            $days = round($ttl / 60 / 60 / 24, 4);
-            mtrace("Time to live (TTL) is $ttl seconds (= $days days)");
-        }
-        // $store->purge_old_entries();
-        // Use reflection to access the private cachestore_file method file_path_for_key
-        $reflection = new \ReflectionClass($store);
-        $filepathmethod = $reflection->getMethod('file_path_for_key');
-        $filepathmethod->setAccessible(true);
-
-        $keys = $store->find_all();
-        $originalcount = count($keys);
-        // Do a get on every key.
-        // The file cache get method should delete keys that are older than ttl but it doesn't...
-        $maxtime = cache::now() - $ttl;
-        foreach ($keys as $key) {
-            // Call the private method
-            $path = $filepathmethod->invoke($store, $key);
-            $filetime = filemtime($path);
-            if ($ttl && $filetime < $maxtime) {
-                $store->delete($key);
+        if (get_config('qtype_coderunner', 'enablegradecache')) {
+            $purger = new cache_purger(usettl:true);
+            $definition = $purger->get_coderunner_cache_definition();
+            $store = $purger->get_file_stores($definition);
+            $ttl = $purger->ttl;
+            if ($ttl) {
+                $days = round($ttl / 60 / 60 / 24, 4);
+                mtrace("Time to live (TTL) is $ttl seconds (= $days days)");
             }
+            // $store->purge_old_entries();
+            // Use reflection to access the private cachestore_file method file_path_for_key
+            $reflection = new \ReflectionClass($store);
+            $filepathmethod = $reflection->getMethod('file_path_for_key');
+            $filepathmethod->setAccessible(true);
+
+            $keys = $store->find_all();
+            $originalcount = count($keys);
+            // Do a get on every key.
+            // The file cache get method should delete keys that are older than ttl but it doesn't...
+            $maxtime = cache::now() - $ttl;
+            foreach ($keys as $key) {
+                // Call the private method
+                $path = $filepathmethod->invoke($store, $key);
+                $filetime = filemtime($path);
+                if ($ttl && $filetime < $maxtime) {
+                    $store->delete($key);
+                }
+            }
+                // $value = $store->get($key);  // Would delete old key if fixed in file store.
+            $remainingkeys = $store->find_all();
+            $newcount = count($remainingkeys);
+            $purgedcount = $originalcount - $newcount;
+            mtrace("Originally found $originalcount keys.");
+            mtrace("$purgedcount keys purged.");
+            mtrace("$newcount keys were too young to die.");
+        } else {
+            mtrace("Grading cache not enabled so not purging.");
         }
-            // $value = $store->get($key);  // Would delete old key if fixed in file store.
-        $remainingkeys = $store->find_all();
-        $newcount = count($remainingkeys);
-        $purgedcount = $originalcount - $newcount;
-        mtrace("Originally found $originalcount keys.");
-        mtrace("$purgedcount keys purged.");
-        mtrace("$newcount keys were too young to die.");
     }
 }
